Postproc[Fix]: fixed problem with voting tests.
This problem was due to the new enhancement for calculating participatio
percentages, that affected both postproc and voting tests.

It's been changed how postproc module creates the results from a list of
dictionaries, to a dictionary consisting in a list of the options (also
dictionaries) with the postproduction attached as an element. And
an element called participation. Example:

'results': [
    { 'option': 'Option 1','number': 1, 'votes': 100000,'postproc': 4 },
    { 'option': 'Option 2','number': 2, 'votes': 80000,'postproc': 3 },
    { 'option': 'Option 3','number': 3, 'votes': 30000,'postproc': 1 },
],
'participation': 100.00,