Visualization[feat]: New import (info in description)
I wasn't sure if add this import since it said in "pyplot's page" that a
dependency was needed to install "pyplot" so I didn't add it while I was
looking for a different way to install the dependency. But I found that
there is a command available inside 'pip install' which lets you install pyplot,
this should also install the needed dependency.